Address CR feedback and suppress warnings
